enhance: 将 swagan 的 generator 改装成工厂模式，通过在generators文件夹下新建 generator js 文件，自动注册generator
