+++ packages/swagen/src/generators/index.js
@@ -1,0 +1,3 @@
+import requireDirectory from "require-directory";
+
+module.exports = requireDirectory(module, { exclude: /lib\.js$/ });
