+++ packages/swagen/src/generators/sdk.js
@@ -1,0 +1,10 @@
+/**
+ * Genereate code for sdk
+ *
+ * @param {*} target
+ * @param {*} swaggerFile
+ * @param {*} name
+ */
+export default function genSDK(target, swaggerFile, name) {
+  // put sdk generation code here
+}
